feat: Expand pretty printer api and implement a basic subset of nodes with tests (#61)
* introduce indent and dedent operator

* some formatting

* add hardline operator

* super basic set of nodes implemented, more tests to be added for broken layouts

* oops

* format and lint

* add a makefile command to also easily lint locally

* lint

* add ability to use tables for alignment

* hide the nest function and only expose the easy to use << and >>

* simplify table impl

* impl not really necessary to expose

* lint

* split table tests

* minor improvements

* add hard-indent and hard-dedent

* dont repeat code

* rangify the clauses

* revert

* start using config

* use config further for structural settings

* higher order bracket helper to prevent mistakes

* swap to config

* lint

* not used

* pass the config to the table

* some test structure improvements

* remove unneeded constructors

* adjust tests

* add explicit constraints

* introduce doc utils

* remove table for now, refactor class order

* wrong comment style

* simplify indent by removing dedent since it really is not needed

* some refactor

* brace init for constructing

* node base should really be an aggregate

* add note

* modify note

* dont use inheritance, lean to composition, get all the aggregate benefits

* refactor tests to use aggregate init

* remove the space helper

* remove the debug printer, add a lot of const

* compound operators

* emplace and not pushback

* .

* dont break declarations

* some niceties

* templates are implicitly inline

* use int for semantic counts and size_t to std::size_t

* get the what

* smoll const

* copilot suggestions, main change and constexpr string_views in tests

* member init with {} for primitive types and remove config member from pretty_printer

* lint

* review

* why not

+++ src/builder/trivia/comment_sink.hpp
@@ -23,7 +23,7 @@ class CommentSink final
             return;
         }
 
-        const auto idx = static_cast<std::size_t>(t->getTokenIndex());
+        const auto idx = t->getTokenIndex();
         if (!used_.insert(idx).second) {
             return; // already added elsewhere
         }
